@@ -1,20 +1,24 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-     <TypeScriptCompileBlocked>true</TypeScriptCompileBlocked>
+    <TargetFramework>net6.0</TargetFramework>
+    <TypeScriptCompileBlocked>true</TypeScriptCompileBlocked>
+    <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
+  
   <ItemGroup>
     <Folder Include="wwwroot\" />
   </ItemGroup>
+
   <ItemGroup>
     <PackageReference Include="BrunoLau.SpaServices" Version="0.8.0" />
-    <PackageReference Include="DotNetify.SignalR" Version="5.0.0" />
+    <PackageReference Include="DotNetify.SignalR" Version="5.2.0" />
     <PackageReference Include="System.Reactive" Version="5.0.0" />
     <PackageReference Include="AspNet.Security.OpenIdConnect.Server" Version="2.0.0" />    
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="server\Services\employees.json" />
   </ItemGroup>
+
   <Target Name="DebugRunWebpack" BeforeTargets="Build" Condition=" '$(Configuration)' == 'Debug' And !Exists('wwwroot\dist') ">
     <!-- Ensure Node.js is installed -->
     <Exec Command="node --version" ContinueOnError="true">
@@ -25,12 +29,16 @@
          a different machine, so rebuild them if not already present. -->
     <Message Importance="high" Text="Performing first-run Webpack build..." />
     <Exec Command="npm install" />
-    <Exec Command="node node_modules/webpack/bin/webpack.js" />
   </Target>
+
+  <Target Name="DebugRunWebpack" BeforeTargets="Build">
+    <Exec Command="npm run build:$(Configuration)" />
+  </Target>  
+
   <Target Name="PublishRunWebpack" AfterTargets="ComputeFilesToPublish">
     <!-- As part of publishing, ensure the JS resources are freshly built in production mode -->
     <Exec Command="npm install" />
-    <Exec Command="node node_modules/webpack/bin/webpack.js --env.prod" />
+    <Exec Command="npm run build:$(Configuration) --env.prod" />
     <!-- Include the newly-built files in the publish output -->
     <ItemGroup>
       <DistFiles Include="wwwroot\**" />
