@@ -1,25 +1,54 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Logging;
+﻿using System.Text;
+using BrunoLau.SpaServices.Webpack;
+using DotNetify;
+using DotNetify.Security;
+using Microsoft.IdentityModel.Tokens;
+using projectName;
 
-namespace projectName
+var builder = WebApplication.CreateBuilder(args);
+
+var services = builder.Services;
+
+// Add OpenID Connect server to produce JWT access tokens.
+services.AddAuthenticationServer();
+
+services.AddSignalR();
+services.AddDotNetify();
+
+services.AddTransient<ILiveDataService, MockLiveDataService>();
+services.AddSingleton<IEmployeeService, EmployeeService>();
+
+var app = builder.Build();
+
+app.UseAuthentication();
+
+app.UseWebSockets();
+app.UseDotNetify(config =>
 {
-   public class Program
+   // Middleware to do authenticate token in incoming request headers.
+   config.UseJwtBearerAuthentication(new TokenValidationParameters
    {
-      public static void Main(string[] args)
-      {
-         BuildWebHost(args).Run();
-      }
+      IssuerSigningKey = new SymmetricSecurityKey(Encoding.ASCII.GetBytes(AuthServer.SecretKey)),
+      ValidateIssuerSigningKey = true,
+      ValidateAudience = false,
+      ValidateIssuer = false,
+      ValidateLifetime = true,
+      ClockSkew = TimeSpan.FromSeconds(0)
+   });
 
-      public static IWebHost BuildWebHost(string[] args) =>
-          WebHost.CreateDefaultBuilder(args)
-              .UseStartup<Startup>()
-              .Build();
-   }
-}
+   // Filter to check whether user has permission to access view models with [Authorize] attribute.
+   config.UseFilter<AuthorizeFilter>();
+});
+
+if (app.Environment.IsDevelopment())
+   app.UseWebpackDevMiddlewareEx(new WebpackDevMiddlewareOptions { HotModuleReplacement = true });
+
+app.UseFileServer();
+app.UseRouting();
+app.UseEndpoints(endpoints =>
+{
+   endpoints.MapHub<DotNetifyHub>("/dotnetify");
+   endpoints.MapFallbackToFile("index.html");
+});
+
+app.Run();
