@@ -1,27 +1,27 @@
-import React from 'react';
-import { IDotnetifyVM, RouteType } from 'dotnetify';
-import { makeStyles } from '@material-ui/core/styles';
-import Card from '@material-ui/core/Card';
-import List from '@material-ui/core/List';
-import ListItem from '@material-ui/core/ListItem';
-import ListItemText from '@material-ui/core/ListItemText';
-import ListItemAvatar from '@material-ui/core/ListItemAvatar';
-import ListItemSecondaryAction from '@material-ui/core/ListItemSecondaryAction';
-import Divider from '@material-ui/core/Divider';
-import Avatar from '@material-ui/core/Avatar';
-import IconButton from '@material-ui/core/IconButton';
-import Menu from '@material-ui/core/Menu';
-import MenuItem from '@material-ui/core/MenuItem';
-import MoreVertIcon from '@material-ui/icons/MoreVert';
-import WallpaperIcon from '@material-ui/icons/Wallpaper';
-import cyan from '@material-ui/core/colors/cyan';
+import React from "react";
+import { IDotnetifyVM, RouteType } from "dotnetify";
+import { makeStyles } from "@material-ui/core/styles";
+import Card from "@material-ui/core/Card";
+import List from "@material-ui/core/List";
+import ListItem from "@material-ui/core/ListItem";
+import ListItemText from "@material-ui/core/ListItemText";
+import ListItemAvatar from "@material-ui/core/ListItemAvatar";
+import ListItemSecondaryAction from "@material-ui/core/ListItemSecondaryAction";
+import Divider from "@material-ui/core/Divider";
+import Avatar from "@material-ui/core/Avatar";
+import IconButton from "@material-ui/core/IconButton";
+import Menu from "@material-ui/core/Menu";
+import MenuItem from "@material-ui/core/MenuItem";
+import MoreVertIcon from "@material-ui/icons/MoreVert";
+import WallpaperIcon from "@material-ui/icons/Wallpaper";
+import cyan from "@material-ui/core/colors/cyan";
 
 const useStyles = makeStyles({
   header: {
     fontSize: 24,
-    fontWeight: 'lighter',
+    fontWeight: "lighter",
     padding: 10,
-    color: 'white',
+    color: "white",
     backgroundColor: cyan[600]
   }
 });
@@ -51,7 +51,7 @@ export default function RecentActivities({ vm, data }: IRecentActivitiesProps) {
       <List>
         {data.map((item, idx) => (
           <React.Fragment key={idx}>
-            <ListItem alignItems='flex-start'>
+            <ListItem alignItems="flex-start">
               <ListItemAvatar>
                 <Avatar>
                   <WallpaperIcon />
@@ -69,7 +69,7 @@ export default function RecentActivities({ vm, data }: IRecentActivitiesProps) {
                 </div>
               </ListItemSecondaryAction>
             </ListItem>
-            <Divider variant='inset' />
+            <Divider variant="inset" />
           </React.Fragment>
         ))}
       </List>
