@@ -1,9 +1,12 @@
-import React from 'react';
-import { Bar } from 'react-chartjs-2';
-import { makeStyles } from '@material-ui/core/styles';
-import Card from '@material-ui/core/Card';
-import pink from '@material-ui/core/colors/pink';
-import globalStyles from '../../styles/styles';
+import React from "react";
+import { Chart as ChartJS, CategoryScale, LinearScale, ArcElement, BarElement, Title, Tooltip, Legend } from "chart.js";
+import { Bar } from "react-chartjs-2";
+import { makeStyles } from "@material-ui/core/styles";
+import Card from "@material-ui/core/Card";
+import pink from "@material-ui/core/colors/pink";
+import globalStyles from "../../styles/styles";
+
+ChartJS.register(CategoryScale, LinearScale, ArcElement, BarElement, Title, Tooltip, Legend);
 
 const useStyles = makeStyles({
   card: {
@@ -11,24 +14,26 @@ const useStyles = makeStyles({
     height: 150
   },
   header: {
-    color: 'white',
+    color: "white",
     backgroundColor: pink[500],
     padding: 10
   },
   body: {
-    marginLeft: 'auto',
-    marginRight: 'auto',
-    width: '95%',
+    marginLeft: "auto",
+    marginRight: "auto",
+    width: "95%",
     height: 85
   }
 });
 
 const chartOptions = {
-  legend: { display: false },
-  tooltips: { enabled: false },
+  plugins: {
+    legend: { display: false },
+    tooltips: { enabled: false }
+  },
   scales: {
-    xAxes: [{ ticks: { fontColor: 'white' }, display: true, gridLines: { display: false } }],
-    yAxes: [{ display: false }]
+    x: { ticks: { color: "white" }, grid: { display: false } },
+    y: { ticks: { display: false } }
   },
   layout: { padding: { bottom: 5 } },
   maintainAspectRatio: false
@@ -52,7 +57,7 @@ export default function ServerUsage(props: IServerUsageProps) {
     ]
   };
 
-  const titleStyle = { ...globalStyles.title }
+  const titleStyle = { ...globalStyles.title };
 
   return (
     <Card className={classes.card}>
