@@ -1,8 +1,11 @@
-import React from 'react';
-import { Line } from 'react-chartjs-2';
-import { makeStyles } from '@material-ui/core/styles';
-import Card from '@material-ui/core/Card';
-import purple from '@material-ui/core/colors/purple';
+import React from "react";
+import { Chart as ChartJS, CategoryScale, LinearScale, PointElement, LineElement, Title, Tooltip, Legend } from "chart.js";
+import { Line } from "react-chartjs-2";
+import { makeStyles } from "@material-ui/core/styles";
+import Card from "@material-ui/core/Card";
+import purple from "@material-ui/core/colors/purple";
+
+ChartJS.register(CategoryScale, LinearScale, PointElement, LineElement, Title, Tooltip, Legend);
 
 const useStyles = makeStyles({
   card: {
@@ -11,19 +14,21 @@ const useStyles = makeStyles({
   },
   header: {
     fontSize: 24,
-    color: 'white',
+    color: "white",
     backgroundColor: purple[600],
     padding: 10
   },
   body: {
     height: 95,
-    padding: '5px 15px 0 15px'
+    padding: "5px 15px 0 15px"
   }
 });
 
 const chartOptions = {
-  legend: { display: false },
-  scales: { xAxes: [{ display: false }], yAxes: [{ display: false }] },
+  plugins: {
+    legend: { display: false }
+  },
+  scales: { xAxis: { ticks: { display: false } }, yAxis: { ticks: { display: false } } },
   layout: { padding: { left: 5, right: 5, top: 5, bottom: 5 } },
   maintainAspectRatio: false
 };
@@ -35,16 +40,15 @@ export interface ITrafficProps {
 export default function Traffic(props: ITrafficProps) {
   const classes = useStyles({});
   const data = {
-    labels: new Array(props.data.length),
+    labels: props.data.map(_ => ""),
     datasets: [
       {
         data: props.data,
         fill: false,
-        backgroundColor: 'white',
-        borderColor: '#8884d8',
+        backgroundColor: "white",
+        borderColor: "#8884d8",
         borderWidth: 2,
-        pointBorderWidth: 2,
-        cubicInterpolationMode: 'monotone'
+        pointBorderWidth: 2
       }
     ]
   };
